+++ server.js
@@ -1,7 +1,7 @@
 const Koa = require('koa');
-const bodyParser = require('koa-bodyparser');
 const ResearchPaperRoutes = require('./routes/ResearchPaper.routes');
 const WorkShopRoutes = require('./routes/WorkShop.routes');
+const FileUploadRoutes = require('./routes/FileUpload.routes');
 const cors = require('@koa/cors');
 const koaBody = require("koa-body");
 
@@ -10,12 +10,11 @@ require('./dal');
 const app = new Koa();
 
 app.use(cors());
-app.use(bodyParser());
-app.use(koaBody({ multipart: true }));
+app.use(koaBody({ multipart: true, json: true }));
 
 app.use(ResearchPaperRoutes.routes()).use(ResearchPaperRoutes.allowedMethods());
-
 app.use(WorkShopRoutes.routes()).use(WorkShopRoutes.allowedMethods());
+app.use(FileUploadRoutes.routes()).use(FileUploadRoutes.allowedMethods());
 
 app.use(ctx =>{
     ctx.body = "Conference Management System Backend"
